+++ src/process/os/linux/cpu_times.rs
@@ -3,11 +3,11 @@ use std::time::Duration;
 use crate::process::ProcessCpuTimes;
 
 pub trait ProcessCpuTimesExt {
-	fn iowait(&self) -> Duration;
+	fn iowait(&self) -> Option<Duration>;
 }
 
 impl ProcessCpuTimesExt for ProcessCpuTimes {
-	fn iowait(&self) -> Duration {
+	fn iowait(&self) -> Option<Duration> {
 		self.iowait
 	}
 }